Clean up tokens and token constants, reduce code duplication, remove unused code.
