+++ jplag.frontend-utils/src/main/java/de/jplag/TokenConstants.java
@@ -2,6 +2,7 @@ package de.jplag;
 
 public interface TokenConstants {
     final static int FILE_END = 0;
+
     /**
      * Used to optionally separate methods from each other with an always marked token
      */
