+++ jplag.frontend.chars/src/main/java/de/jplag/chars/CharToken.java
@@ -3,17 +3,18 @@ package de.jplag.chars;
 import de.jplag.Token;
 
 public class CharToken extends Token {
-    private static final long serialVersionUID = 1L;
 
     private int index;
+    private Character character;
 
-    public CharToken(int type, String file, int index, Parser parser) {
+    public CharToken(int type, String file, int index, Character character, Parser parser) {
         super(type, file, -1);
         this.index = index;
+        this.character = character;
     }
 
     public CharToken(int type, String file, Parser parser) {
-        this(type, file, -1, parser);
+        this(type, file, -1, null, parser);
     }
 
     @Override
@@ -34,5 +35,28 @@ public class CharToken extends Token {
     @Override
     public int getIndex() {
         return index;
+    }
+
+    @Override
+    public void setLine(int line) {
+        // Do nothing!
+    }
+
+    @Override
+    public void setColumn(int column) {
+        // Do nothing!
+    }
+
+    @Override
+    public void setLength(int length) {
+        // Do nothing!
+    }
+
+    @Override
+    protected String type2string() {
+        if (character == null) {
+            return "FILE_END";
+        }
+        return character.toString();
     }
 }
