+++ jplag.frontend.chars/src/main/java/de/jplag/chars/Language.java
@@ -3,7 +3,6 @@ package de.jplag.chars;
 import java.io.File;
 
 import de.jplag.ErrorConsumer;
-import de.jplag.Token;
 import de.jplag.TokenList;
 
 /*
@@ -73,10 +72,5 @@ public class Language implements de.jplag.Language {
     @Override
     public int numberOfTokens() {
         return 36;
-    }
-
-    @Override
-    public String type2string(int type) {
-        return Token.type2string(type);
     }
 }
