+++ jplag.frontend.cpp/src/main/java/de/jplag/cpp/CPPTokenConstants.java
@@ -3,11 +3,6 @@ package de.jplag.cpp;
 import de.jplag.TokenConstants;
 
 public interface CPPTokenConstants extends TokenConstants {
-    final static int FILE_END = 0;
-    // Used to optionally separate methods from each other
-    // with an always marked token
-    final static int SEPARATOR_TOKEN = 1;
-
     final static int C_BLOCK_BEGIN = 2;
     final static int C_BLOCK_END = 3;
     final static int C_SCOPE = 4;
