+++ jplag.frontend.csharp-1.2/src/main/java/de/jplag/csharp/CSharpTokenConstants.java
@@ -3,12 +3,6 @@ package de.jplag.csharp;
 import de.jplag.TokenConstants;
 
 public interface CSharpTokenConstants extends TokenConstants {
-    final static int FILE_END = 0;
-
-    // Used to optionally separate methods from each other
-    // with an always marked token
-    final static int SEPARATOR_TOKEN = 1;
-
     final static int _INVOCATION = 2;
     final static int _OBJECT_CREATION = 3;
     final static int _ARRAY_CREATION = 4;
