+++ jplag.frontend.java/src/main/java/de/jplag/java/JavaToken.java
@@ -3,44 +3,13 @@ package de.jplag.java;
 import de.jplag.Token;
 
 public class JavaToken extends Token implements JavaTokenConstants {
-    private static final long serialVersionUID = -383581430479870696L;
-    private int line, column, length;
 
     public JavaToken(int type, String file, int line, int col, int length) {
         super(type, file, line, col, length);
     }
 
     @Override
-    public int getLine() {
-        return line;
-    }
-
-    @Override
-    public int getColumn() {
-        return column;
-    }
-
-    @Override
-    public int getLength() {
-        return length;
-    }
-
-    @Override
-    public void setLine(int line) {
-        this.line = line;
-    }
-
-    @Override
-    public void setColumn(int column) {
-        this.column = column;
-    }
-
-    @Override
-    public void setLength(int length) {
-        this.length = length;
-    }
-
-    public static String type2string(int type) {
+    protected String type2string() {
         switch (type) {
             case JavaTokenConstants.FILE_END:
                 return "********";
