+++ jplag.frontend.java/src/main/java/de/jplag/java/JavaTokenConstants.java
@@ -3,11 +3,8 @@ package de.jplag.java;
 import de.jplag.TokenConstants;
 
 public interface JavaTokenConstants extends TokenConstants {
-    final static int FILE_END = 0;
 
-    // check used to mark tokens that are actually generated by the adapter
-    final static int SEPARATOR_TOKEN = 1;// @formatter:off
-
+    // @formatter:off
     final static int J_PACKAGE = 2;				// check
     final static int J_IMPORT = 3;				// check
     final static int J_CLASS_BEGIN = 4;			// check
