+++ jplag.frontend.scheme/src/main/java/de/jplag/scheme/Language.java
@@ -70,9 +70,4 @@ public class Language implements de.jplag.Language {
     public int numberOfTokens() {
         return SchemeTokenConstants.NUM_DIFF_TOKENS;
     }
-
-    @Override
-    public String type2string(int type) {
-        return SchemeToken.type2string(type);
-    }
 }
