+++ jplag.frontend.text/src/main/java/de/jplag/text/TextToken.java
@@ -19,9 +19,6 @@ public class TextToken extends Token {
         return serial.intValue();
     }
 
-    // ///////////////////// END OF STATIC MEMBERS
-
-    private int line, column, length;
     private String text;
 
     public TextToken(int type, String file, Parser parser) {
@@ -34,37 +31,12 @@ public class TextToken extends Token {
         this.text = text.toLowerCase();
     }
 
-    @Override
-    public int getLine() {
-        return line;
-    }
-
-    @Override
-    public int getColumn() {
-        return column;
-    }
-
-    @Override
-    public int getLength() {
-        return length;
-    }
-
-    @Override
-    public void setLine(int line) {
-        this.line = line;
-    }
-
-    @Override
-    public void setColumn(int column) {
-        this.column = column;
-    }
-
-    @Override
-    public void setLength(int length) {
-        this.length = length;
-    }
-
     public String getText() {
         return this.text;
+    }
+
+    @Override
+    protected String type2string() {
+        return getText();
     }
 }
